内源检视问题修改 Signed-off-by: zuojiangjiang <zuojiangjiang@huawei.com>

+++ frameworks/innerkitsimpl/distributeddatafwk/src/ikvstore_observer.cpp
@@ -98,9 +98,8 @@ void KvStoreObserverProxy::OnChange(const ChangeNotification &changeNotification
         ZLOGE("Write ChangeNotification buffer size to parcel failed.");
         return;
     }
-    ZLOGD("I(%lld) U(%lld) D(%lld) T(%lld)", static_cast<long long>(insertBufferSize),
-          static_cast<long long>(updateBufferSize), static_cast<long long>(deleteBufferSize),
-          static_cast<long long>(totalBufferSize));
+    ZLOGD("I(%" PRId64") U(%" PRId64") D(%" PRId64") T(%" PRId64")",
+          insertBufferSize, updateBufferSize, deleteBufferSize, totalBufferSize);
     if (totalBufferSize < Constant::SWITCH_RAW_DATA_SIZE) {
         if (!data.WriteParcelable(&changeNotification)) {
             ZLOGW("Write ChangeNotification to parcel failed.");
